backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
pymongo==4.6.*
redis==5.0.*
PyJWT==2.8.*
bcrypt==4.1.*
python-decouple==3.8
django-ratelimit==4.1.*
django-cors-headers==4.3.*
pytest==7.4.*
pytest-django==4.7.*
factory-boy==3.3.*
